moved updateVotedPost in Posthelpers.ts
